pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python_template"
version = "0.1.0"
description = "Template"
authors = ["DataForGood"]
license = " MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
plotly = "^5.22.0"
matplotlib = "^3.9.0"
dash = "^2.17.1"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
dash-mantine-components = "^0.14"
dash-iconify = "^0.1.2"
jinja2 = "^3.1"
jupyterlab = "^4.2.5"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.2.0"
tox = "^4.4.8"

[tool.ruff]
# Same as Black.
line-length = 96
target-version = "py310"

# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "B"]
ignore = ["E501"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.mccabe]
max-complexity = 10
